requirements.txt
langchain==0.0.214
langchain-openai==0.1.1
langchain-community==0.0.31
openai==1.16.2
chromadb==0.4.24
tiktoken==0.6.0
psycopg2-binary==2.9.9
flask==3.0.3
numpy==1.25.0
scikit-learn==1.4.1.post1
matplotlib==3.8.4
japanize_matplotlib==1.1.3
nltk==3.8.1
janome

